Debugging of NewDoit for 1d.
The azimuth integration of scattering matrix for 1D was wrong.
"Old" DOIT and NewDOIT agree within the order of 1-2 percent.
The reason for the difference is the changed gas absorption
handling. Whereas the "old" Doit calculate the gas absorption
for every p_path step, NewDOIT calculates the gas absorption
only on the p_grd and interpolates them on the p_path_step.

Possible solution, the gas absorption can be calculated on
additional and finer p_grid, which is then used for the
interpolation.