Updated on forward scattering correction
Correction now uses interpolation of order 2 for the forward peak.
It seems that the correction reduces the polarization a little bit.
As we use only energy conservation for the adjustment of the
extinction matrix, we do not adjust the 3,4 term of the extinction
matrix. Furthermore, the correction assumes at least azimuthal
random orientation. For now, it is apllicable only for 1D.