Implemented a higher order intergartion scheme for RT step
A higher order integration scheme was implemented, which
replaces the old 1st order integration schemen, which has
difficulties in strong scattering cases without much
absorption. So far, the scheme seems to work but tests
against RT4 and DISORT are NEEDED!

Furthermore a bug in a method of the propagation matrix
cloass was corrected.