Revise ppath_step calculation to be completely local.
* So far ppath_step calculation were done within the cloudbox but with respect to
the global p_, lat_ and lon_grid, z_, vmr_ and t_field. Now, the ppath_step calculation
is local. This means, it is relative to the cloudbox domain. Therefore, we had for now to
hardwire the ppath_stepGeometric into calculation of the cloud propagation paths, because
to use the agenda mechanism we need to modify several different methods that do not belong to
NewDoit.